+++ src/thvm.c
@@ -60,6 +60,7 @@ static void init_default_ctx_scalars(TContext *ctx) {
 #include "term/new_fvr.c"
 #include "term/new_bri.c"
 #include "term/new_ann.c"
+#include "term/new_pri.c"
 
 // === heap/ ===
 #include "heap/alloc.c"
@@ -188,6 +189,7 @@ static void init_default_ctx_scalars(TContext *ctx) {
 #include "interact/dup_num.c"
 #include "interact/dup_any.c"
 #include "interact/app_bri.c"
+#include "interact/app_pri.c"
 #include "interact/ann_lam.c"
 #include "interact/ann_bri.c"
 #include "interact/dup_bri.c"